Move session form to own component
