+++ apps/store/src/app/debugger/components/CreateSessionForm/CreateSessionForm.css.ts
@@ -1,0 +1,6 @@
+import { style } from '@vanilla-extract/css'
+
+export const wrapper = style({
+  maxWidth: '400px',
+  margin: 'auto',
+})
