+++ apps/store/src/app/debugger/components/CreateSessionForm/CreateSessionForm.tsx
@@ -3,7 +3,8 @@
 import { useEffect, useState } from 'react'
 import { Button, Space } from 'ui'
 import { TextField } from '@/components/TextField/TextField'
-import { create } from './actions'
+import { create } from '../../actions'
+import { wrapper } from './CreateSessionForm.css'
 
 const HEDVIG_DEBUGGER_SSN = 'hedvig:debugger-ssn'
 
@@ -16,17 +17,20 @@ export const CreateSessionForm = () => {
   }, [])
 
   return (
-    <form action={create}>
-      <Space y={0.25}>
-        <TextField
-          label="YYYYMMDDXXXX"
-          key={defaultSsn}
-          defaultValue={defaultSsn}
-          autoFocus
-          name="ssn"
-        />
-        <Button>Create session</Button>
-      </Space>
-    </form>
+    <div className={wrapper}>
+      <form action={create}>
+        <Space y={0.25}>
+          <TextField
+            label="YYYYMMDDXXXX"
+            key={defaultSsn}
+            defaultValue={defaultSsn}
+            autoFocus
+            required
+            name="ssn"
+          />
+          <Button>Create session</Button>
+        </Space>
+      </form>
+    </div>
   )
 }
