+++ apps/store/src/app/debugger/page.tsx
@@ -1,28 +1,11 @@
-import Head from 'next/head'
-import { CreateSessionForm } from './CreateSessionForm'
+import { CreateSessionForm } from './components/CreateSessionForm/CreateSessionForm'
 
 const Page = () => {
-  return (
-    <div style={wrapperStyle}>
-      <Head>
-        <meta name="robots" content="noindex,nofollow" />
-      </Head>
-      <CreateSessionForm />
-    </div>
-  )
+  return <CreateSessionForm />
 }
 
 export default Page
 
 export const metadata = {
   robots: 'noindex, nofollow',
-}
-
-const wrapperStyle = {
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  height: '100%',
-  maxWidth: '20rem',
-  marginInline: 'auto',
 }
